+++ EleCfitsioWrapper/EleCfitsioWrapper/impl/HduWrapper.hpp
@@ -19,7 +19,9 @@
 
 #if defined(_ELECFITSIOWRAPPER_HDUWRAPPER_IMPL) || defined(CHECK_QUALITY)
 
+#include "EleCfitsioWrapper/BintableWrapper.h"
 #include "EleCfitsioWrapper/HduWrapper.h"
+#include "EleCfitsioWrapper/ImageWrapper.h"
 
 #include <utility> // index_sequence, make_index_sequence
 
@@ -47,7 +49,7 @@ template <typename... TInfos>
 void initBintableExtension(fitsfile* fptr, const std::string& name, const TInfos&... infos) {
   constexpr long ncols = sizeof...(TInfos);
   Fits::String::CStrArray colName {infos.name...};
-  Fits::String::CStrArray colFormat {TypeCode<typename TInfos::Value>::tform(infos.repeatCount)...};
+  Fits::String::CStrArray colFormat {TypeCode<typename TInfos::Value>::tform(infos.repeatCount())...};
   Fits::String::CStrArray colUnit {infos.unit...};
   int status = 0;
   fits_create_tbl(fptr, BINARY_TBL, 0, ncols, colName.data(), colFormat.data(), colUnit.data(), name.c_str(), &status);
@@ -84,7 +86,7 @@ void assignBintableExtension(fitsfile* fptr, const std::string& name, const TCol
   constexpr long columnCount = 1;
   std::string colName = column.info().name;
   char* cName = &colName[0];
-  std::string colFormat = TypeCode<typename TColumn::Value>::tform(column.info().repeatCount);
+  std::string colFormat = TypeCode<typename TColumn::Value>::tform(column.info().repeatCount());
   char* cFormat = &colFormat[0];
   std::string colUnit = column.info().unit;
   char* cUnit = &colUnit[0];