+++ EleFitsData/EleFitsData/ColumnInfo.h
@@ -55,54 +55,72 @@ namespace Fits {
 /**
  * @ingroup bintable_data_classes
  * @brief Column informations, i.e. name, unit, entry shape and value type.
+ * @tparam T The value type
+ * @tparam N The dimension (number of axes per entry)
  * @details
- * Binary tables can be seen as a sequence of columns or rows, made of consecutive entries.
+ * Binary tables can be seen as a sequence of columns made of consecutive entries.
  * Entry values are not necessarily simple types:
  * they can be either string, scalar, vector or multidimensional.
  * Yet, all the entries of one column have the same properties.
+ * The column informations consist in:
+ * - The value type as template parameter `T`;
+ * - The dimension (number of axes) as template parameter `N`;
+ * - The column `name`;
+ * - The column `unit`;
+ * - The entry `shape` (see below).
  * 
+ * The number of values per entry, named repeat count, is the shape size.
  * Here is a table which summarizes the properties of each column type
  * -- and more details follow.
  * 
  * <table class="fieldtable">
- * <tr><th>Type<th>`T`<th>`N`<th>`repeatCount`<th>`elementCount`<th>`shape`
- * <tr><td>String<td>`std::string`<td>1<td>> max number of characters<td>1<td>Undefined
- * <tr><td>Scalar<td>Not `std::string`<td>1<td>1<td>1<td>Undefined
- * <tr><td>Vector<td>Not `std::string`<td>1<td>> 1<td> = `repeatCount`<td>Undefinerd
- * <tr><td>Multidimensional<td>Not `std::string`<td>-1 or > 1<td>&ge; shape size<td>= `repeatCount`<td>Unconstrained
+ * <tr><th>Category<th>`T`<th>`N`<th>`repeatCount()`<th>`elementCount()`<th>`shape`
+ * <tr><td>%String<td>`std::string`<td>1<td>> max number of characters<td>1<td>`{repeatCount()}`
+ * <tr><td>Scalar<td>Not `std::string`<td>1<td>1<td>1<td>`{1}`
+ * <tr><td>Vector<td>Not `std::string`<td>1<td>> 1<td> = repeat count<td>`{repeatCount()}`
+ * <tr><td>Multidimensional<td>Not `std::string`<td>-1 or > 1<td>= shape size<td>= shape size<td>Unconstrained
  * </table>
  * 
- * In the case of vector and multidimensional columns, each entry contains `repeatCount` values.
- * Here is an example of a 4-row table with a scalar column, a vector column and a string column:
+ * For string columns, the element count differs from the repeat count,
+ * in that the element count is the number of `std::string` objects stored in the column data container,
+ * while the repeat count is the number of characters allocated to each string in the Fits file.
+ * This results in `elementCount()` = 1 and `repeatCount()` > 1
+ * (multidimensional string columns are not supported).
+ * 
+ * @par_example
+ * Here is an example of a 4-row table with each column category:
+ * 
  * <table class="fieldtable">
- * <tr><th>Row<th>`repeatCount` = 1<th>`repeatCount` = 3<th>`repeatCount` = 6
- * <tr><td>0<td>00<td>00, 01, 02<td>`"ZERO"`
- * <tr><td>1<td>10<td>10, 11, 12<td>`"ONE"`
- * <tr><td>2<td>20<td>20, 21, 22<td>`"TWO"`
- * <tr><td>3<td>30<td>30, 31, 32<td>`"THREE"`
+ * <tr><th>Row<th>%String<th>Scalar<th>Vector<th>Multidim
+ * <tr><td>0<td>`"ZERO"`<td>0<td>00, 01, 02<td>000, 001, 002<br>010, 011, 012
+ * <tr><td>1<td>`"ONE"`<td>1<td>10, 11, 12<td>100, 101, 102<br>110, 111, 112
+ * <tr><td>2<td>`"TWO"`<td>2<td>20, 21, 22<td>200, 201, 202<br>210, 211, 212
+ * <tr><td>3<td>`"THREE"`<td>3<td>30, 31, 32<td>300, 301, 302<br>310, 311, 312
  * </table>
+ * 
  * For performance, the values are stored sequentially in a 1D array as follows:
  * \code
- * int repeat1[] = {00, 10, 20, 30};
- * int repeat3[] = {00, 01, 02, 10, 11, 12, 20, 21, 22, 30, 31, 32};
- * \endcode
- *
- * The only exception to this is string columns,
- * for which the data array is a simple array of `std::string`s:
- * \code
- * std::string data[] = {"ZERO", "ONE", "TWO", "THREE"};
- * \endcode
- * but the repeat count should be at least 6 (beware of the null terminating character).
+ * ColumnInfo<std::string> stringInfo("String", "", 6);
+ * std::string stringData[] = {"ZERO", "ONE", "TWO", "THREE"};
  * 
- * For convenience, we distinguish:
- * - repeat count, which is defined by the Fits standard;
- * - element count, which is the number of elements of type `T` in the container;
- * - element count per row, which is the element count divided by the row count.
+ * ColumnInfo<int> scalarInfo("Scalar");
+ * int scalarData[] = {0, 1, 2, 3};
  * 
- * Therefore, for `T = std::string`, the element count per row is 1,
- * while the repeat count should be large enough to hold all characters.
- * For `T != std::string`, the element count per row is the repeat count.
- *
+ * ColumnInfo<int> vectorInfo("Vector", "", 3);
+ * int vectorData[] = {
+ *     00, 01, 02,
+ *     10, 11, 12,
+ *     20, 21, 22,
+ *     30, 31, 32};
+ * 
+ * ColumnInfo<int, 2> multidimInfo("Multidim", "", {3, 2});
+ * int multidimData[] = {
+ *     000, 001, 002, 010, 011, 012,
+ *     100, 101, 102, 110, 111, 112,
+ *     200, 201, 202, 210, 211, 212,
+ *     300, 301, 302, 310, 311, 312};
+ * \endcode
+ * 
  * @note
  * Since the values are stored sequentially even for vector columns,
  * a scalar column can be "fold" into a vector column by
@@ -122,9 +140,8 @@ namespace Fits {
  * Therefore, users are encouraged to consider the repeat count as a meaningful value,
  * rather than as an optimization trick.
  * 
- * @see ColumnInfo< T, 1 > for unidimensional cases.
+ * @see `Column`
  * @see \ref optim
- * @see \ref data_classes
  */
 template <typename T, long N = 1>
 struct ColumnInfo {
@@ -148,8 +165,7 @@ struct ColumnInfo {
    * The entry shape is deduced from the repeat count
    * (first axis' lenght is the repeat count, others are set to 1).
    */
-  ColumnInfo(std::string n = "", std::string u = "", long r = 1) :
-      name(n), unit(u), repeatCount(r), shape(Position<N>::one()) {
+  ColumnInfo(std::string n = "", std::string u = "", long r = 1) : name(n), unit(u), shape(Position<N>::one()) {
     shape[0] = r;
   }
 
@@ -161,8 +177,7 @@ struct ColumnInfo {
    * @details
    * The repeat count is deduced from the shape.
    */
-  ColumnInfo(std::string n, std::string u, Position<N> s) :
-      name(n), unit(u), repeatCount(shapeSize(s)), shape(std::move(s)) {}
+  ColumnInfo(std::string n, std::string u, Position<N> s) : name(n), unit(u), shape(std::move(s)) {}
 
   /**
    * @brief The column name.
@@ -175,20 +190,17 @@ struct ColumnInfo {
   std::string unit;
 
   /**
-   * @brief Get the repeat count.
-   */
-  long repeatCount;
-
-  /**
    * @brief The shape of one entry.
    */
   Position<N> shape;
 
   /**
+   * @brief Get the repeat count.
+   */
+  long repeatCount() const;
+
+  /**
    * @brief Get the number of elements per entry.
-   * @details
-   * This is generally the repeat count,
-   * except for string columns for which it always equals 1.
    */
   long elementCount() const;
 };